Sun, Feb 11, 2024, 9:34 PM -08:00

+++ filter-less/helpers.c
@@ -9,7 +9,7 @@ void grayscale(int height, int width, RGBTRIPLE image[height][width])
         for (int column = 0; column < width; column++)
         {
             RGBTRIPLE pixel = image[row][column];
-            int pixel_brightness = (pixel.rgbtBlue + pixel.rgbtGreen + pixel.rgbtRed)/3;
+            int pixel_brightness = (int) (pixel.rgbtBlue + pixel.rgbtGreen + pixel.rgbtRed)/3;
             image[row][column].rgbtBlue = pixel_brightness;
             image[row][column].rgbtGreen = pixel_brightness;
             image[row][column].rgbtRed = pixel_brightness;
